Start incorporating Router
