+++ package.json
@@ -8,6 +8,7 @@
     "@testing-library/user-event": "^13.5.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "5.0.1",
     "sass": "^1.54.5",
     "web-vitals": "^2.1.4"
