+++ src/Movie-Card/Movie-Card.js
@@ -1,7 +1,8 @@
 import React from 'react'
+import { Link } from 'react-router-dom';
 
-const MovieCard = ({ id, title, poster, switchView }) => {
-    return <img key={id} src={poster} id={id} name={title} alt={title} className='movie-card' height='300' width='200' onClick={() => switchView(id)}/>
+const MovieCard = ({ id, title, poster }) => {
+    return <Link to={'/' + id}><img key={id} src={poster} id={id} name={title} alt={title} className='movie-card' height='300' width='200'/></Link>
 }
 
 export default MovieCard;
