+++ src/Navbar/Navbar.scss
@@ -33,6 +33,10 @@
     }
 }
 
+.navbar-home {
+    text-decoration: none;
+}
+
 .search-by-title-input {
     margin-right: 20px;
     padding: 7px 15px;
